All examples in FIELDSPEC now generated from API
All examples start with API framgment in examples.py

These are then used to generate the API code and the raw spec JSON.
This is ordered and formatted along with the yaml for the example.

Each example has a number of iterations and any extra commands line
pipes or extra args to run against the generated spec.

The output of N iterations is captured and stored as part of the example.

This should allow regression testing in the future and increased consistancy
as well as the ability to easily add new examples that can be compiled
and validated, reducing the chances of an invalid example being documented.